% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.xi_acf}
\alias{autoplot.xi_acf}
\title{Plot method for xi_acf objects}
\usage{
\method{autoplot}{xi_acf}(object, ...)
}
\arguments{
\item{object}{An object of class \code{xi_acf}.}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Plot method for xi_acf objects
}
