% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.xi_ccf}
\alias{autoplot.xi_ccf}
\title{Plot method for xi_ccf objects}
\usage{
\method{autoplot}{xi_ccf}(object, ...)
}
\arguments{
\item{object}{An object of class \code{xi_ccf}.}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Plot method for xi_ccf objects
}
