% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.xi_matrix}
\alias{autoplot.xi_matrix}
\title{Plot method for xi_matrix objects}
\usage{
\method{autoplot}{xi_matrix}(object, ...)
}
\arguments{
\item{object}{An object of class \code{xi_matrix}.}

\item{...}{Additional arguments passed to other methods.}
}
\description{
Plot method for xi_matrix objects
}
