% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.xi_test.R
\name{autoplot.xi_test}
\alias{autoplot.xi_test}
\title{Plot Xi-ACF Comparison}
\usage{
\method{autoplot}{xi_test}(object, ...)
}
\arguments{
\item{object}{An object of class "xi_test".}

\item{...}{Additional arguments (ignored).}
}
\description{
Plot Xi-ACF Comparison
}
