% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_xi_acf_maxstat_cpp}
\alias{compute_xi_acf_maxstat_cpp}
\title{Compute empirical Xi-ACF and Max-Statistic null distribution}
\usage{
compute_xi_acf_maxstat_cpp(x, max_lag, n_surr, max_iter = 100L)
}
\description{
Compute empirical Xi-ACF and Max-Statistic null distribution
}
\keyword{internal}
