% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_xi_ccf_cpp}
\alias{compute_xi_ccf_cpp}
\title{Compute Lagged Xi Cross-Correlation with MIAAFT Surrogates}
\usage{
compute_xi_ccf_cpp(x, y, max_lag, n_surr)
}
\arguments{
\item{x}{A numeric vector (Variable 1)}

\item{y}{A numeric vector (Variable 2)}

\item{max_lag}{The maximum lag to compute (computes from -max_lag to +max_lag)}

\item{n_surr}{Number of MIAAFT surrogates to generate}
}
\value{
A list containing lags, original xi values, and surrogate xi values.
}
\description{
Compute Lagged Xi Cross-Correlation with MIAAFT Surrogates
}
