% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_xi_ccf_maxstat_cpp}
\alias{compute_xi_ccf_maxstat_cpp}
\title{Compute empirical Xi-CCF and Max-Statistic null distribution}
\usage{
compute_xi_ccf_maxstat_cpp(
  x,
  y,
  max_lag,
  n_surr,
  max_iter = 100L,
  both_directions = TRUE
)
}
\description{
Compute empirical Xi-CCF and Max-Statistic null distribution
}
\keyword{internal}
