% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_xi_ccf_miaaft}
\alias{compute_xi_ccf_miaaft}
\title{Compute MIAAFT-based Directional Xi-CCF}
\usage{
compute_xi_ccf_miaaft(x, y, max_lag, n_surr)
}
\arguments{
\item{x}{First time series (numeric vector, potential cause)}

\item{y}{Second time series (numeric vector, potential effect)}

\item{max_lag}{Maximum positive lag to evaluate}

\item{n_surr}{Number of surrogate datasets to generate}
}
\value{
A list containing forward (X leads Y) and backward (Y leads X) Xi coefficients and surrogates.
}
\description{
Compute MIAAFT-based Directional Xi-CCF
}
