% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_xi_lags}
\alias{compute_xi_lags}
\title{Compute Xi Statistics for Lags (Core Engine)}
\usage{
compute_xi_lags(x, max_lag, n_surr)
}
\arguments{
\item{x}{Numeric vector (time series).}

\item{max_lag}{Maximum lag to compute.}

\item{n_surr}{Number of surrogate datasets to generate.}
}
\value{
A list containing `xi_original` and `xi_surrogates`.
}
\description{
Calculates Chatterjee's Xi for multiple lags and generates IAAFT surrogates.
}
