% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_xi_matrix_miaaft}
\alias{compute_xi_matrix_miaaft}
\title{Compute Pairwise Directional Xi-CCF for a Multivariate Matrix}
\usage{
compute_xi_matrix_miaaft(x, max_lag, n_surr)
}
\arguments{
\item{x}{A numeric matrix (rows = time, cols = variables).}

\item{max_lag}{An integer specifying the maximum positive lag.}

\item{n_surr}{An integer specifying the number of surrogate datasets.}
}
\value{
A list containing flat vectors for lead/lag variable indices, lags, 
        original Xi values, and a matrix of surrogate Xi values.
}
\description{
Compute Pairwise Directional Xi-CCF for a Multivariate Matrix
}
