% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_miaaft_surrogate_cpp}
\alias{generate_miaaft_surrogate_cpp}
\title{Generate a Single MIAAFT Surrogate Matrix}
\usage{
generate_miaaft_surrogate_cpp(x, max_iter = 100L)
}
\arguments{
\item{x}{A numeric matrix (rows = time, cols = variables).}

\item{max_iter}{An integer specifying the maximum number of iterations.}
}
\value{
A numeric matrix representing the generated MIAAFT surrogate.
}
\description{
Generate a Single MIAAFT Surrogate Matrix
}
