% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_acf.R
\name{print.xi_acf}
\alias{print.xi_acf}
\title{Print method for xi_acf objects}
\usage{
\method{print}{xi_acf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"xi_acf"}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the original object.
}
\description{
Print method for xi_acf objects
}
