% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_test.R
\name{print.xi_test}
\alias{print.xi_test}
\title{Print method for xi_test objects}
\usage{
\method{print}{xi_test}(x, ...)
}
\arguments{
\item{x}{An object of class "xi_test".}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for xi_test objects
}
