% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{surrogate_iaaft_cpp}
\alias{surrogate_iaaft_cpp}
\title{Generate Multiple IAAFT Surrogates (Univariate)}
\usage{
surrogate_iaaft_cpp(x, n_surr, max_iter = 100L)
}
\arguments{
\item{x}{A numeric vector.}

\item{n_surr}{Number of surrogates to generate.}

\item{max_iter}{Maximum iterations for IAAFT.}
}
\value{
A matrix of surrogates (N x n_surr).
}
\description{
Generate Multiple IAAFT Surrogates (Univariate)
}
