% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{surrogate_miaaft_cpp}
\alias{surrogate_miaaft_cpp}
\title{Generate Multiple MIAAFT Surrogates (3D Array / Cube)}
\usage{
surrogate_miaaft_cpp(X, n_surr, max_iter = 100L)
}
\arguments{
\item{X}{A numeric matrix (N x p).}

\item{n_surr}{Number of surrogates to generate.}

\item{max_iter}{Maximum iterations for MIAAFT.}
}
\value{
A 3D array (arma::cube) of dimensions N x p x n_surr.
}
\description{
Generate Multiple MIAAFT Surrogates (3D Array / Cube)
}
