% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_fft_cpp}
\alias{test_fft_cpp}
\title{Test FFT and IFFT in C++}
\usage{
test_fft_cpp(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list containing the original, amplitudes, phases, and reconstructed vector.
}
\description{
Test FFT and IFFT in C++
}
