% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xi_coefficient}
\alias{xi_coefficient}
\title{Compute Chatterjee's Xi coefficient (Exported to R)}
\usage{
xi_coefficient(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
The Xi coefficient.
}
\description{
Compute Chatterjee's Xi coefficient (Exported to R)
}
