% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_test.R
\name{xi_test}
\alias{xi_test}
\title{Xi-ACF Test for Time Series}
\usage{
xi_test(x, max_lag = 20, n_surr = 100)
}
\arguments{
\item{x}{Numeric vector (time series).}

\item{max_lag}{Integer. Maximum lag to compute.}

\item{n_surr}{Integer. Number of surrogates for IAAFT test.}
}
\value{
An object of class "xi_test".
}
\description{
Calculates Chatterjee's Xi and Standard ACF with significance thresholds.
}
