% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.xi_acf.R
\name{autoplot.xi_acf}
\alias{autoplot.xi_acf}
\title{Plot Xi-ACF Comparison}
\usage{
\method{autoplot}{xi_acf}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"xi_acf"}.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A \code{ggplot} object representing the correlogram.
}
\description{
Visualizes the comparison between the standard linear Autocorrelation Function (ACF)
and the non-linear Chatterjee's Xi coefficient, including their respective significance thresholds.
}
