% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.xi_matrix.R
\name{autoplot.xi_matrix}
\alias{autoplot.xi_matrix}
\title{Plot Multivariate Xi-Correlogram Matrix}
\usage{
\method{autoplot}{xi_matrix}(object, ...)
}
\arguments{
\item{object}{An object of class \code{xi_matrix}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A \code{ggplot} object.
}
\description{
Visualizes the result of a multivariate Xi-matrix analysis using a facet grid.
Rows represent the leading (predictor) variable, and columns represent the lagging (response) variable.
}
