% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_xi_acf}
\alias{extract_xi_acf}
\title{Extract Individual Xi-ACF from a Multivariate Xi-Matrix}
\usage{
extract_xi_acf(object, var, x_raw = NULL)
}
\arguments{
\item{object}{An object of class \code{xi_matrix}.}

\item{var}{A character string specifying the variable name.}

\item{x_raw}{Optional. The original data to calculate linear ACF.}
}
\value{
An object of class \code{xi_acf}.
}
\description{
Extract Individual Xi-ACF from a Multivariate Xi-Matrix
}
