% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_xi_acf}
\alias{extract_xi_acf}
\title{Extract Univariate Xi-ACF from a Multivariate Xi-Matrix}
\usage{
extract_xi_acf(obj, var, ...)
}
\arguments{
\item{obj}{An object of class \code{xi_matrix}.}

\item{var}{A character string specifying the variable name to extract.}

\item{...}{Additional arguments passed to xi_acf.}
}
\value{
An object of class \code{xi_acf}.
}
\description{
Extract Univariate Xi-ACF from a Multivariate Xi-Matrix
}
