% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_xi_acf}
\alias{extract_xi_acf}
\title{Extract Individual Xi-ACF from Xi-Matrix}
\usage{
extract_xi_acf(object, var, x_raw = NULL)
}
\arguments{
\item{object}{An object of class \code{xi_matrix}.}

\item{var}{A character string specifying the variable name to extract.}

\item{x_raw}{Optional. The original multivariate data (matrix or data.frame) used to
compute the matrix. If provided, standard linear ACF will be re-calculated.}
}
\value{
An object of class \code{xi_acf}.
}
\description{
Extracts the autocorrelation results for a specific variable from a \code{xi_matrix} object
and converts it into a \code{xi_acf} S3 object.
}
