% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_xi_ccf}
\alias{extract_xi_ccf}
\title{Extract Bivariate Xi-CCF from a Multivariate Xi-Matrix}
\usage{
extract_xi_ccf(obj, var_x, var_y, ...)
}
\arguments{
\item{obj}{An object of class \code{xi_matrix}.}

\item{var_x}{A character string specifying the lead variable.}

\item{var_y}{A character string specifying the lag variable.}

\item{...}{Additional arguments passed to xi_ccf.}
}
\value{
An object of class \code{xi_ccf}.
}
\description{
Extract Bivariate Xi-CCF from a Multivariate Xi-Matrix
}
