% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_xi_ccf}
\alias{extract_xi_ccf}
\title{Extract Pairwise Xi-CCF from a Multivariate Xi-Matrix}
\usage{
extract_xi_ccf(object, var_x, var_y, x_raw = NULL)
}
\arguments{
\item{object}{An object of class \code{xi_matrix}.}

\item{var_x}{Variable X name.}

\item{var_y}{Variable Y name.}

\item{x_raw}{Optional. The original data to calculate linear CCF.}
}
\value{
An object of class \code{xi_ccf}.
}
\description{
Extract Pairwise Xi-CCF from a Multivariate Xi-Matrix
}
