% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_ccf.R
\name{print.xi_ccf}
\alias{print.xi_ccf}
\title{Print method for xi_ccf objects}
\usage{
\method{print}{xi_ccf}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"xi_ccf"}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the original object.
}
\description{
Print method for xi_ccf objects
}
