% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_matrix.R
\name{print.xi_matrix}
\alias{print.xi_matrix}
\title{Print method for xi_matrix}
\usage{
\method{print}{xi_matrix}(x, ...)
}
\arguments{
\item{x}{An object of class \code{xi_matrix}.}

\item{...}{Additional arguments passed to print.}
}
\value{
The original object \code{x} invisibly. Called primarily for its side effect of printing the matrix to the console.
}
\description{
Print method for xi_matrix
}
