% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_acf.R
\name{xi_acf}
\alias{xi_acf}
\alias{xi_test}
\title{Xi-ACF Test for Time Series}
\usage{
xi_acf(x, max_lag = 20, n_surr = 100)

xi_test(x, max_lag = 20, n_surr = 100)
}
\arguments{
\item{x}{A numeric vector representing the time series data.}

\item{max_lag}{An integer specifying the maximum number of lags to compute.}

\item{n_surr}{An integer specifying the number of surrogate datasets to generate for the IAAFT test.}
}
\value{
An object of class \code{"xi_acf"} containing the computed statistics and metadata.
}
\description{
Calculates Chatterjee's Xi and the standard Autocorrelation Function (ACF)
along with their respective significance thresholds.
}
