% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xi_coefficient}
\alias{xi_coefficient}
\title{Calculate Chatterjee's Rank Correlation Coefficient (Xi)}
\usage{
xi_coefficient(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\value{
A numeric scalar representing the Chatterjee's Xi coefficient.
}
\description{
Computes Chatterjee's rank correlation coefficient (Xi) between two numeric vectors.
Ties are broken uniformly at random to ensure strict inequalities.
}
