% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xi_coefficient}
\alias{xi_coefficient}
\title{Calculate Chatterjee's Xi Coefficient}
\usage{
xi_coefficient(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\value{
Double value of Xi
}
\description{
Calculate Chatterjee's Xi Coefficient
}
