% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_matrix.R
\name{xi_matrix}
\alias{xi_matrix}
\title{Multivariate Xi-Correlogram Matrix}
\usage{
xi_matrix(x, max_lag = 10, n_surr = 399, sig_level = 0.05, max_iter = 100, ...)
}
\arguments{
\item{x}{A numeric matrix or data.frame containing the multivariate time series (columns = variables).}

\item{max_lag}{An integer specifying the maximum positive lag to compute. Default is 10.}

\item{n_surr}{An integer specifying the number of MIAAFT surrogate datasets. Default is 399.}

\item{sig_level}{A numeric value between 0 and 1 specifying the significance level. Default is 0.05.}

\item{max_iter}{An integer specifying the maximum iterations for the MIAAFT algorithm. Default is 100.}

\item{...}{Additional arguments.}
}
\value{
An S3 object of class \code{xi_matrix} containing a tidy data frame of pairwise results.
}
\description{
Computes the pairwise directional Chatterjee's Xi coefficient for a multivariate time series.
}
