% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{xiacf-deprecated}
\alias{xiacf-deprecated}
\alias{xi_test}
\alias{generate_iaaft_surrogate}
\alias{generate_miaaft_surrogates}
\alias{generate_miaaft_surrogate_cpp}
\alias{run_rolling_xi_analysis}
\alias{compute_xi_acf_iaaft}
\alias{compute_xi_ccf_miaaft}
\alias{compute_xi_matrix_miaaft}
\title{Deprecated functions in xiacf}
\usage{
xi_test(x, max_lag = 10, n_surr = 399, sig_level = 0.95, max_iter = 100, ...)

generate_iaaft_surrogate(...)

generate_miaaft_surrogates(...)

generate_miaaft_surrogate_cpp(...)

run_rolling_xi_analysis(...)

compute_xi_acf_iaaft(...)

compute_xi_ccf_miaaft(...)

compute_xi_matrix_miaaft(...)
}
\arguments{
\item{x}{A numeric vector or matrix depending on the function.}

\item{max_lag}{An integer specifying the maximum lag.}

\item{n_surr}{An integer specifying the number of surrogate datasets.}

\item{sig_level}{A numeric value specifying the significance or confidence level.}

\item{max_iter}{An integer specifying the maximum iterations.}

\item{...}{Additional arguments passed to the updated functions.}
}
\description{
These functions are provided for backward compatibility with
older versions of xiacf and will be removed in future releases.
}
